+++ apps/frontend/app/admin/course/[courseId]/user/_components/InviteByCode.tsx
@@ -1,7 +1,3 @@
-'use client'
-
-import { ErrorMessage } from '@/app/admin/_components/ErrorMessage'
-import { Modal } from '@/components/Modal'
 import {
   AlertDialog,
   AlertDialogAction,
@@ -13,296 +9,27 @@ import {
   AlertDialogTitle
 } from '@/components/shadcn/alert-dialog'
 import { Button } from '@/components/shadcn/button'
-import { Input } from '@/components/shadcn/input'
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue
-} from '@/components/shadcn/select'
 import { Switch } from '@/components/shadcn/switch'
 import { CREATE_WHITE_LIST, DELETE_WHITE_LIST } from '@/graphql/course/mutation'
 import { GET_COURSE, GET_WHITE_LIST } from '@/graphql/course/queries'
-import {
-  INVITE_USER,
-  ISSUE_INVITATION,
-  REVOKE_INVITATION
-} from '@/graphql/user/mutation'
-import { GET_GROUP_MEMBERS } from '@/graphql/user/queries'
-import { fetcherWithAuth } from '@/libs/utils'
-import type { MemberRole } from '@/types/type'
-import { useApolloClient, useMutation, useQuery } from '@apollo/client'
-import { valibotResolver } from '@hookform/resolvers/valibot'
-import { useParams } from 'next/navigation'
-import { useCallback, useEffect, useState } from 'react'
+import { ISSUE_INVITATION, REVOKE_INVITATION } from '@/graphql/user/mutation'
+import { useMutation, useQuery } from '@apollo/client'
+import { useEffect, useState } from 'react'
 import { CSVLink } from 'react-csv'
-import { useForm, type SubmitHandler } from 'react-hook-form'
-import { HiMiniPlusCircle } from 'react-icons/hi2'
+import { useForm } from 'react-hook-form'
 import { IoCloudUpload, IoCopyOutline } from 'react-icons/io5'
-import { MdOutlineEmail } from 'react-icons/md'
 import { toast } from 'sonner'
 import * as XLSX from 'xlsx'
-import { findUserSchema, inviteUserSchema } from '../_libs/schema'
-
-export function InviteButton() {
-  const client = useApolloClient()
-  const { courseId } = useParams<{ courseId: string }>()
-
-  const [isAlertDialogOpen, setIsAlertDialogOpen] = useState(false)
-
-  // const handleOpenChange = (isOpen: boolean) => {
-  //   setIsAlertDialogOpen(isOpen)
-  //   if (!isOpen) {
-  //     refetchGroupMembers()
-  //   }
-  // }
-
-  const refetchGroupMembers = () => {
-    client.refetchQueries({
-      include: [GET_GROUP_MEMBERS]
-    })
-  }
-
-  return (
-    <Modal
-      size="lg"
-      type="custom"
-      title="Invite Member"
-      // open={isAlertDialogOpen}
-      onOpenChange={refetchGroupMembers}
-      trigger={
-        <Button type="button" variant="default" className="w-[120px]">
-          <HiMiniPlusCircle className="mr-2 h-5 w-5" />
-          <span className="text-lg">Invite</span>
-        </Button>
-      }
-    >
-      <div className="flex flex-1 flex-col gap-6 overflow-hidden px-4">
-        <InviteByCode
-          courseId={courseId}
-          isAlertDialogOpen={isAlertDialogOpen}
-        />
-        <InviteManually courseId={courseId} />
-      </div>
-    </Modal>
-  )
-}
-
-interface InviteUserInput {
-  groupId: number
-  userId: number
-  isGroupLeader: boolean
-}
-
-interface FindUserInput {
-  email: string
-}
-
-interface InviteManuallyProps {
-  courseId: string
-}
-
-interface UserInfo {
-  userName: string
-  id: number
-}
-
-type InvitedUserDisplay = {
-  email: string
-  role: 'Instructor' | 'Student'
-}
-
-function InviteManually({ courseId }: InviteManuallyProps) {
-  const roles: MemberRole[] = ['Instructor', 'Student']
-  const [userId, setUserId] = useState(0)
-
-  const [invitedList, setInvitedList] = useState<InvitedUserDisplay[]>([])
-  const [inviteUser] = useMutation(INVITE_USER)
-
-  const onFind: SubmitHandler<FindUserInput> = async (data) => {
-    try {
-      const res = await fetcherWithAuth('user/email', {
-        searchParams: { email: data.email }
-      })
-
-      if (!res.ok) {
-        toast.error('Failed to find user')
-        return
-      }
-
-      const userInfo: UserInfo = await res.json()
-      setUserId(userInfo.id)
-    } catch (err) {
-      console.error(err)
-      toast.error('Unexpected error occurred')
-    }
-  }
-
-  const onInvite: SubmitHandler<InviteUserInput> = useCallback(
-    async (data) => {
-      if (!userId) {
-        toast.error('No user selected to invite')
-        return
-      }
-
-      try {
-        const result = await inviteUser({
-          variables: {
-            groupId: Number(courseId),
-            isGroupLeader: data.isGroupLeader,
-            userId
-          }
-        })
-
-        setInvitedList((prevList) => [
-          ...prevList,
-          {
-            email: result.data?.inviteUser.user.email ?? '',
-            role: result.data?.inviteUser.isGroupLeader
-              ? 'Instructor'
-              : 'Student'
-          }
-        ])
-
-        toast.success('Invited Successfully!', {
-          style: {
-            background: '#F0F8FF',
-            color: '#0973DC',
-            borderRadius: '1000px',
-            border: '1px solid rgba(255, 255, 255, 0.10)',
-            maxWidth: '200px'
-          },
-          closeButton: false
-        })
-      } catch {
-        toast.error('Failed to invite user')
-      }
-    },
-    [inviteUser, courseId, userId, setInvitedList]
-  )
-
-  const {
-    register: findRegister,
-    handleSubmit: findHandleSubmit,
-    formState: { errors: findErrors }
-  } = useForm<FindUserInput>({
-    resolver: valibotResolver(findUserSchema)
-  })
-
-  const {
-    watch: inviteWatch,
-    handleSubmit: inviteHandleSubmit,
-    setValue: inviteSetValue,
-    formState: { errors: inviteErrors }
-  } = useForm<InviteUserInput>({
-    resolver: valibotResolver(inviteUserSchema),
-    defaultValues: {
-      userId,
-      groupId: Number(courseId),
-      isGroupLeader: false
-    }
-  })
-
-  useEffect(() => {
-    if (userId !== 0) {
-      inviteHandleSubmit(onInvite)()
-    }
-  }, [inviteHandleSubmit, onInvite, userId])
-
-  return (
-    <div className="flex flex-1 flex-col overflow-hidden">
-      <form
-        onSubmit={(event) => {
-          event.preventDefault()
-          findHandleSubmit(onFind)()
-        }}
-        aria-label="Invite user"
-        className="flex flex-col gap-4"
-      >
-        <span className="text-base font-bold">Invite Manually</span>
-
-        <div className="flex w-full max-w-[500px] flex-col gap-4">
-          <div className="flex h-[44px] items-center rounded-full border border-gray-300 px-4 py-2">
-            <MdOutlineEmail className="h-5 w-5 text-gray-400" />
-
-            <Input
-              id="email"
-              {...findRegister('email')}
-              placeholder="Email Address"
-              className="flex-1 border-none !bg-transparent pl-2 text-sm placeholder:text-gray-400 autofill:!bg-transparent focus:outline-none focus:ring-0 focus-visible:border-none focus-visible:outline-none focus-visible:ring-0"
-            />
-
-            <Select
-              value={inviteWatch('isGroupLeader') ? 'Instructor' : 'Student'}
-              onValueChange={(value) => {
-                inviteSetValue('isGroupLeader', value === 'Instructor')
-              }}
-            >
-              <SelectTrigger className="w-auto min-w-[80px] border-none bg-transparent text-sm text-gray-500 focus:outline-none">
-                <SelectValue placeholder="Student" />
-              </SelectTrigger>
-              <SelectContent className="rounded-lg bg-white shadow-md">
-                {roles.map((role) => (
-                  <SelectItem key={role} value={role}>
-                    {role}
-                  </SelectItem>
-                ))}
-              </SelectContent>
-            </Select>
-          </div>
-
-          <Button
-            type="submit"
-            className="bg-primary hover:bg-primary-strong h-[44px] w-full rounded-full text-sm font-semibold text-white"
-          >
-            Invite
-          </Button>
-        </div>
-
-        {findErrors.email && (
-          <ErrorMessage message={findErrors.email.message} />
-        )}
-        {inviteErrors.groupId && (
-          <ErrorMessage message={inviteErrors.groupId.message} />
-        )}
-        {inviteErrors.userId && (
-          <ErrorMessage message={inviteErrors.userId.message} />
-        )}
-        {inviteErrors.isGroupLeader && (
-          <ErrorMessage message={inviteErrors.isGroupLeader.message} />
-        )}
-      </form>
-
-      {invitedList.length === 0 && <div className="mt-[60px]" />}
-
-      {invitedList.length > 0 && (
-        <div className="flex flex-1 flex-col gap-2 overflow-y-auto pt-2">
-          {invitedList.map((user) => (
-            <div
-              key={user.email}
-              className="flex h-[42px] w-full items-start justify-between rounded-full bg-gray-100 px-[24px] py-[10px]"
-            >
-              <span className="text-sm text-gray-800">{user.email}</span>
-              <span className="text-sm text-gray-400">{user.role}</span>
-            </div>
-          ))}
-        </div>
-      )}
-    </div>
-  )
-}
 
 interface InviteByCodeProps {
   courseId: string
-  isAlertDialogOpen: boolean
 }
 
 interface InvitationCodeInput {
   issueInvitation: string
 }
 
-function InviteByCode({ courseId, isAlertDialogOpen }: InviteByCodeProps) {
+export function InviteByCode({ courseId }: InviteByCodeProps) {
   const [isInviteByCodeEnabled, setIsInviteByCodeEnabled] = useState(false)
   const [isApprovalRequired, setIsApprovalRequired] = useState(false)
   const [issueInvitation] = useMutation(ISSUE_INVITATION)
@@ -346,11 +73,9 @@ function InviteByCode({ courseId, isAlertDialogOpen }: InviteByCodeProps) {
   })
 
   useEffect(() => {
-    if (isAlertDialogOpen) {
-      refetchWhiteList()
-      refetchInvitationCode()
-    }
-  }, [isAlertDialogOpen, refetchInvitationCode, refetchWhiteList])
+    refetchWhiteList()
+    refetchInvitationCode()
+  }, [refetchInvitationCode, refetchWhiteList])
 
   const { getValues, reset } = useForm<InvitationCodeInput>()
